+++ src/app/core/core.models.ts
@@ -1,3 +1,6 @@
+import { Observable } from 'rxjs';
+import { UrlTree } from '@angular/router';
+
 export interface ConfigModel {
   id: string;
   login: string;
@@ -6,4 +9,21 @@ export interface ConfigModel {
   firstName?: string;
   lastName?: string;
   phone?: string;
+}
+
+export enum Role {
+  User = 'User',
+  Admin = 'Admin',
+}
+
+export class User {
+  role!: Role;
+}
+
+export interface ICanComponentDeactivate {
+  canDeactivate: () =>
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree;
 }