+++ src/app/core/services/local-storage/local-storage.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-@Injectable()
+@Injectable({ providedIn: 'root' })
 export class LocalStorageService {
   setItem(key: string, item: any): void {
     localStorage.setItem(key, JSON.stringify(item));