+++ src/app/orders/orders-routing.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ProcessOrderComponent } from './components/process-order/process-order.component';
+
+const routes: Routes = [{ path: '', component: ProcessOrderComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class OrdersRoutingModule {}