+++ src/app/products/products-routing.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ProductListComponent } from './components/product-list/product-list.component';
+import { ProductViewComponent } from './components/product-view/product-view.component';
+import { firstComponentItem } from './components/first/first.component';
+
+const routes: Routes = [
+  {
+    path: 'product-list',
+    component: ProductListComponent,
+  },
+  {
+    path: 'product/0',
+    component: ProductViewComponent,
+    pathMatch: 'full',
+    data: firstComponentItem,
+  },
+  {
+    path: 'product/:productID',
+    component: ProductViewComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ProductsRoutingModule {}